[Kernel]: Bug fixes in argument string passing [Kernel]: Bug fixes in process data structure [Kernel]: Bug fixes in system call handler, where it checks [Kernel]: Added PipeFS for adding Named Pipes, previous pipes module is changed to named pipe, no traditional pipe is supported system call number for out of bound errors [DeodhaiAudio]: implemented deodhai audio server with mixing various audio streams with mono and stereo channels, adding special features like panning, gain control, and little compression to 16 bit samples

+++ Drivers/Sound/IntelHDA/ihda.cpp
@@ -329,7 +329,8 @@ void HDAHandler(size_t v, void* p) {
 			uint32_t pos = dma[4] & 0xffffffff;
 			pos /= BUFFER_SIZE;
 			pos %= BDL_SIZE;
-			AuSoundGetBlock((uint64_t*)(_hdaudio->sample_buffer + pos * BUFFER_SIZE));
+			AuSoundGetBlock((uint64_t*)(_hdaudio->sample_buffer + 
+				static_cast<uint64_t>(pos) * BUFFER_SIZE));
 		}
 
 	}
